CRUD cloud Account and file refactor, with mongoGrid CRUD working and the Read and upload to drive also working

+++ final-project-file-server/uploader/src/database/repositories/cloudStorageAccount.repository.ts
@@ -5,6 +5,7 @@ import { ICloudStorageAccountRepository } from '../../types/ICloudStorageAccount
 import { CloudStorageAccountMapper } from '../../mappers/cloudStorageAccount.mapper'
 import { CloudStorageAccountDTO } from '../../dto/cloudStorageAccount.dto'
 import { CloudStorageAccountEntity } from '../entities/cloudStorageAccount.entity'
+import { ObjectId } from 'mongodb'
 
 export class CloudStorageAccountRepository implements ICloudStorageAccountRepository {
   private readonly repository: Repository<CloudStorageAccountEntity>
@@ -15,35 +16,54 @@ export class CloudStorageAccountRepository implements ICloudStorageAccountReposi
 
   createCloudStorageAccount = async (cloudStorageAccount: CloudStorageAccountDTO) => {
     const createCloudStorageAccount = await this.repository.save(CloudStorageAccountMapper.toEntity(cloudStorageAccount))
-    return CloudStorageAccountMapper.toDTO(createCloudStorageAccount)
+    return createCloudStorageAccount
   }
 
   readCloudStorageAccount = async (cloudStorageAccountId: string) => {
-    const foundCloudStorageAccount = await this.repository.findOneBy({ _id: cloudStorageAccountId })
+    if (!cloudStorageAccountId) throw new Error('cloudStorageAccountId not provided')
+
+    const id: ObjectId = new ObjectId(cloudStorageAccountId)
+
+    const foundCloudStorageAccount = await this.repository.findOneBy({ _id: id })
 
     if (!foundCloudStorageAccount) {
       throw new Error(`Cloud Storage Account with id ${cloudStorageAccountId} not found`)
     }
 
-    return CloudStorageAccountMapper.toDTO(foundCloudStorageAccount)
+    return foundCloudStorageAccount
   }
 
-  updateCloudStorageAccount = async (cloudStorageAccount: CloudStorageAccountDTO) => {
-    const { cloudStorageId } = cloudStorageAccount
-    const updatedCloudStorageAccount = await this.repository.update({ _id: cloudStorageId }, CloudStorageAccountMapper.toEntity(cloudStorageAccount))
+  readCloudStorageAccounts = async () => {
+    const foundCloudStorageAccounts = await this.repository.find()
+
+    if (!foundCloudStorageAccounts) {
+      throw new Error('Cloud Storage Accounts not found')
+    }
+
+    return foundCloudStorageAccounts
+  }
+
+  updateCloudStorageAccount = async (cloudStorageAccountId: string, cloudStorageAccount: CloudStorageAccountDTO) => {
+    if (!cloudStorageAccountId) throw new Error('cloudStorageAccountId not provided')
+
+    const id: ObjectId = new ObjectId(cloudStorageAccountId)
+    const updatedCloudStorageAccount = await this.repository.update({ _id: id }, CloudStorageAccountMapper.toEntity(cloudStorageAccount))
 
     const didMatch = updatedCloudStorageAccount.raw.matchedCount > 0
 
     if (!didMatch) {
-      throw new Error(`Cloud Storage Account ${cloudStorageId}, matchedCount equal to 0`)
+      throw new Error(`Cloud Storage Account ${cloudStorageAccountId}, matchedCount equal to 0`)
     }
 
     return didMatch
   }
 
   deleteCloudStorageAccount = async (cloudStorageAccountId: string) => {
+    if (!cloudStorageAccountId) throw new Error('cloudStorageAccountId not provided')
+
+    const id: ObjectId = new ObjectId(cloudStorageAccountId)
     const options: FindManyOptions<CloudStorageAccountEntity> = {
-      where: { _id: cloudStorageAccountId }
+      where: { _id: id }
     }
     const cloudStorage = await this.repository.find(options)
     const removeResponse = await this.repository.remove(cloudStorage)
